+++ examples/bode.rs
@@ -2,7 +2,11 @@ extern crate automatica;
 
 use automatica::plots::bode::BodePlot;
 use automatica::transfer_function::Tf;
-use automatica::{polynomial::Poly, Decibel, Eval};
+use automatica::{
+    polynomial::Poly,
+    units::{Decibel, RadiantsPerSecond},
+    Eval,
+};
 
 use num_complex::Complex;
 
@@ -14,10 +18,10 @@ fn main() {
     let c = tf.eval(&Complex::new(0., 1.));
     println!("{}\n{}dB, {}°", c, c.norm().to_db(), c.arg().to_degrees());
 
-    let b = tf.bode(0.1, 10.0, 0.1);
+    let b = tf.bode(RadiantsPerSecond(0.1), RadiantsPerSecond(10.0), 0.1);
     for g in b.into_db_deg() {
         println!(
-            "f: {:.3} rad, m: {:.3} dB, f: {:.1} °",
+            "f: {:.3} rad/s, m: {:.3} dB, f: {:.1} °",
             g.angular_frequency(),
             g.magnitude(),
             g.phase()