+++ examples/linear_system.rs
@@ -1,8 +1,6 @@
 extern crate automatica;
 
-use automatica::linear_system::Ss;
-use automatica::transfer_function::TfMatrix;
-use automatica::Eval;
+use automatica::{linear_system::Ss, transfer_function::TfMatrix, units::Seconds, Eval};
 
 use num_complex::Complex;
 
@@ -21,11 +19,11 @@ fn main() {
     println!("\nStep response:");
     let step = |_| vec![1.0];
 
-    let rk2 = sys.rk2(step, &[0., 0.], 0.1, 150);
+    let rk2 = sys.rk2(step, &[0., 0.], Seconds(0.1), 150);
     println!("rk2 stationary values: {:?}", rk2.last().unwrap());
     // Change to 'true' to print the result
     if false {
-        for i in sys.rk2(step, &[0., 0.], 0.1, 150) {
+        for i in sys.rk2(step, &[0., 0.], Seconds(0.1), 150) {
             println!(
                 "{};{};{};{};{}",
                 i.time(),
@@ -37,11 +35,11 @@ fn main() {
         }
     }
 
-    let rk4 = sys.rk4(step, &[0., 0.], 0.1, 150);
+    let rk4 = sys.rk4(step, &[0., 0.], Seconds(0.1), 150);
     println!("rk4 stationary values: {:?}", rk4.last().unwrap());
     // Change to 'true' to print the result
     if false {
-        for i in sys.rk4(step, &[0., 0.], 0.1, 150) {
+        for i in sys.rk4(step, &[0., 0.], Seconds(0.1), 150) {
             println!(
                 "{};{};{};{};{}",
                 i.time(),
@@ -53,11 +51,11 @@ fn main() {
         }
     }
 
-    let rkf45 = sys.rkf45(step, &[0., 0.], 0.1, 16., 1e-4);
+    let rkf45 = sys.rkf45(step, &[0., 0.], Seconds(0.1), Seconds(16.), 1e-4);
     println!("rkf45 stationary values: {:?}", rkf45.last().unwrap());
     // Change to 'true' to print the result
     if false {
-        for i in sys.rkf45(step, &[0., 0.], 0.1, 16., 1e-4) {
+        for i in sys.rkf45(step, &[0., 0.], Seconds(0.1), Seconds(16.), 1e-4) {
             println!(
                 "{};{};{};{};{};{}",
                 i.time(),
@@ -70,11 +68,11 @@ fn main() {
         }
     }
 
-    let radau = sys.radau(step, &[0., 0.], 0.1, 150, 1e-4);
+    let radau = sys.radau(step, &[0., 0.], Seconds(0.1), 150, 1e-4);
     println!("radau stationary values: {:?}", radau.last().unwrap());
     // Change to 'true' to print the result
     if false {
-        for i in sys.radau(step, &[0., 0.], 0.1, 150, 1e-4) {
+        for i in sys.radau(step, &[0., 0.], Seconds(0.1), 150, 1e-4) {
             println!(
                 "{:>4.1};{:>9.6};{:>9.6};{:>9.6};{:>9.6}",
                 i.time(),