+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+## [0.5.0] - 2019-09-08
+### Added
+- Discretization of transfer functions
+- Units of measurement
+### Changed
+- Use typed unit of measurement instead of primitive types
+
 ## [0.4.1] - 2019-09-01
 ### Added
 - Documentation links for discrete system.